WIP: broke saxs_math out of saxs_fit module, moved parameterize_spectrum to saxs_regressor module
